+++ manage.py
@@ -8,19 +8,15 @@ import cad
 import imagejops
 import heatmap
 import metrics
+import align
+import utility
 
-output_folder = r"E:\greg\Run4"
+output_folder = r"E:\greg\Results\Run8"
 cad_folder = r"C:\temporary work folder gsk35\UIUCMxD\CAD"
 images_folder = r"C:\temporary work folder gsk35\UIUCMxD\Images"
 
-Path(output_folder).mkdir(parents=True, exist_ok=True) # exist_ok=False to prevent overwriting any directories by accident
-
 csv_directory = (r"C:\temporary work folder gsk35\UIUCMxD\mxd_258key.csv")
 parts_df = pd.read_csv(csv_directory)
-
-# choose which operations will be done and which parameters they will use
-# do_steps_dict = imagejops.DEFAULT_DO_STEPS_DICT
-# steps_parameters_dict = imagejops.DEFAULT_STEPS_PARAMETERS_DICT
 
 do_steps = {
     "subtract_background" : True,
@@ -40,43 +36,71 @@ steps_parameters = {
     "threshold" : 40,
 }
 
-sift_align_argument = imagejops.DEFAULT_ALIGN_ARGUMENT
+Path(output_folder).mkdir(parents=True, exist_ok=True)
+Path(os.path.join(output_folder, "cads")).mkdir(parents=True, exist_ok=True)
+Path(os.path.join(output_folder, "aligned_scans")).mkdir(parents=True, exist_ok=True)
+Path(os.path.join(output_folder, "extra_pixels")).mkdir(parents=True, exist_ok=True)
+Path(os.path.join(output_folder, "missing_pixels")).mkdir(parents=True, exist_ok=True)
+Path(os.path.join(output_folder, "heatmaps")).mkdir(parents=True, exist_ok=True)
+Path(os.path.join(output_folder, "steps")).mkdir(parents=True, exist_ok=True)
 
 # write json file with details of operations and parameters
 metadata_dict = {
     "do_steps_dict" : do_steps,
     "steps_parameters_dict" : steps_parameters,
-    "sift_align_argument" : sift_align_argument
 }
-with open(os.path.join(output_folder, "metadata.json"), "x") as fp:
-    json.dump(metadata_dict, fp)
+if not os.path.isfile(os.path.join(output_folder, "metadata.json")):
+    with open(os.path.join(output_folder, "metadata.json"), "x") as fp:
+        json.dump(metadata_dict, fp)
+else:
+    print("WARNING: a metadata file already exists. Metadata will be saved in metadata_overwrite.json. It is at risk of being overwritten next time this code is run.")
+    with open(os.path.join(output_folder, "metadata_overwrite.json"), "w") as fp:
+        json.dump(metadata_dict, fp)
 
 rows = parts_df.iterrows()
 # rows = parts_df.loc[parts_df["serial"] == "GUIDE-0121-0143"].iterrows()
+# rows = parts_df[parts_df.index.isin([235,236])].iterrows()
 
 for idx, row in tqdm(rows, total=parts_df.shape[0]):
-    # make destination folder for the two files
-    dest_folder_path = os.path.join(output_folder, row["serial"])
-    Path(dest_folder_path).mkdir(parents=True, exist_ok=False)
+    try:
+        width, height = utility.get_file_image_dimensions(os.path.join(images_folder, row["img_name"]))
 
-    # process and write scan file to destination folder
-    width, height = imagejops.process_scan(os.path.join(images_folder, row["img_name"]),
-                                           os.path.join(dest_folder_path, "image1.tif"),
-                                           do_steps,
-                                           steps_parameters)
+        # process and write cad file to destination folder
+        img_cad = cad.cad_to_img(os.path.join(cad_folder, row["mockingbird_file"]),
+                                os.path.join(output_folder, "cads", row["serial"] + "_cad.tif"),
+                                width, height)
 
-    # process and write cad file to destination folder
-    img_cad = cad.cad_to_img(os.path.join(cad_folder, row["mockingbird_file"]),
-                             os.path.join(dest_folder_path, "image0.tif"),
-                             width, height)
-    
-    # align cad & scan
-    Path(os.path.join(dest_folder_path, "aligned")).mkdir(parents=True, exist_ok=False)
-    imagejops.align(dest_folder_path, os.path.join(dest_folder_path, "aligned"), sift_align_argument)
+        # align scan to cad
+        align.align_ccorr(
+            os.path.join(output_folder, "cads", row["serial"] + "_cad.tif"),
+            os.path.join(images_folder, row["img_name"]),
+            {
+                "aligned" : os.path.join(output_folder, "aligned_scans", row["serial"] + "_aligned.tif"),
+                "extra_pixels" : os.path.join(output_folder, "extra_pixels", row["serial"] + "_extra.tif"),
+                "missing_pixels" : os.path.join(output_folder, "missing_pixels", row["serial"] + "_missing.tif")
+            }
+        )
 
-    # heatmap
-    heatmap.heatmap(os.path.join(dest_folder_path, "image0.tif"),
-                    os.path.join(dest_folder_path, "aligned", "Aligned 2 of 20001.tif"),
-                    os.path.join(dest_folder_path, "heatmap.tif"))
-    
-metrics.get_directory_metrics(output_folder)
+        # process and write scan file to destination folder
+        imagejops.process_scan(os.path.join(output_folder, "aligned_scans", row["serial"] + "_aligned.tif"),
+                               os.path.join(output_folder, "steps", row["serial"] + "_processed.tif"),
+                               do_steps,
+                               steps_parameters,
+                               save_steps=False)
+        
+        # processing introduces some artifacts around previous crop. remove these by cropping again with a slightly smaller kernel.
+        align.crop_to_inflated_cad(os.path.join(output_folder, "cads", row["serial"] + "_cad.tif"),
+                                   os.path.join(output_folder, "steps", row["serial"] + "_processed.tif"),
+                                   os.path.join(output_folder, "steps", row["serial"] + "_processed.tif"),
+                                   kernel_size=275,
+                                   overwrite=True)
+
+        # heatmap
+        heatmap.heatmap(os.path.join(output_folder, "cads", row["serial"] + "_cad.tif"),
+                        os.path.join(output_folder, "steps", row["serial"] + "_processed.tif"),
+                        os.path.join(output_folder, "heatmaps", row["serial"] + "_heatmap.tif"))
+        
+    except OSError as e:
+        print(f"Skipped {row['serial']} due to OSError. Error was: {e}")
+        
+metrics.get_directory_metrics(os.path.join(output_folder, "heatmaps"), output_folder)
